add IPv6 support
The `myip` query parameter can now contain one or two IP-addresses.
If only one address is provided it can be either an IPv4 or an IPv6
address. If two addresses are provided they must be comma separated
and one address must be an IPv4 address and the other an IPv6 address
in any order.

Both the `A` and the `AAAA` record sets are updated at the same time and
kept in sync. This also implies: When `myip` contains only one address
the resource record set for that address is created or updated and the
resource record set for the other address type is deleted if it exists.

The structure of the configuration in the code has been simplified by
eliminating the `record` structure and moving the `ttl` parameter up
into the host level.