remove (broken) support for multiple hostnames per request
The DynDNS protocol allows a comma separated list of names in the `hostname`
query parameter. The old code splitted the parameter and checked the
configuration but effectively updated at most one hostname because

    any(r53_upsert(host, ..., ip) for host in hosts)

will stop calling `r53_upsert` after that function returned `True` once.

The DynDNS protocol also requires a `good` or `nochg` line per host in
the response observing the order given in the query parameter.

Fixing all this is more complicated than it's worth.