Allow empty components in `myip` or even empty `myip`
As a consequence `myip=` will delete existing `A` and `AAAA` records for the
given host.

The problem: The client uses a pattern like `myip=<ipaddr>,<ip6addr>`.
But it can happen that only an IPv4 address is assigned initially leading
to `myip=123.123.123.123,` which the old code rejected due to the empty
component. After some seconds the IPv6 address is assigned but the client is
scared by the previous `badagent` response.